feat: total paid expenses, move in lib/budget
